# 8281168: Micro-optimize VarForm.getMemberName for interpreter