--- a/src/java.base/share/classes/java/lang/invoke/VarForm.java
+++ b/src/java.base/share/classes/java/lang/invoke/VarForm.java
@@ -109,6 +109,8 @@ final class VarForm {
 
     @ForceInline
     final MemberName getMemberName(int mode) {
+        // Can be simplified by calling getMemberNameOrNull, but written in this
+        // form to improve interpreter/coldpath performance.
         MemberName mn = memberName_table[mode];
         if (mn == null) {
             mn = resolveMemberName(mode);