@cl4es 2022-02-08
> Thanks. Any more reviews needed, or I am pushing?

I think you're good to push. I suspect some might question making the code ever so less obvious/maintainable on behalf of interpreter performance, but I think it's well-motivated since startup speed is where MHs/VHs continue to feel a bit lackluster.