@mlchung 2022-02-08
This change looks okay.    One biggest cold startup overhead we measured for JEP 416 is due to the overhead of spinning and loading classes of MH/VH.      This micro-optimization focuses on the performance of VH invocation.    Do you see  class spinning and loading from your benchmark?   Maybe in the generic invocation case?