@shipilev 2022-02-08
> This change looks okay. One biggest cold startup overhead we measured for JEP 416 is due to the overhead of spinning and loading classes of MH/VH. This micro-optimization focuses on the performance of VH invocation. Do you see class spinning and loading from your benchmark? Maybe in the generic invocation case?

Not really, I am trying to see that `-Xint` does not have easy to fix impediments to cold performance. In this benchmark, bootstrapping stuff had already happened at warmup, it measures the overheads of doing all the MH/VH without the optimizing compiler assistance.